chore(config): move the styleguidist prompt to another file
