+++ styleguide.config.js
@@ -1,37 +1,10 @@
 'use strict';
 const { version } = require('./package');
 const path = require('path');
-const kleur = require('kleur');
-const webpackDevServerUtils = require('react-dev-utils/WebpackDevServerUtils');
+const { printServerInstructions } = require('./styleguidist.print');
 
 // Make react-scripts get the eslint config
 process.env.EXTEND_ESLINT = 'true';
-
-const printServerInstructions = (config) => {
-  const urls = webpackDevServerUtils.prepareUrls('http', config.serverHost, config.serverPort);
-  console.log(`${kleur.cyan().bold(`
-/££££££££
-|_____ ££
-     /££/   /££££££   /££££££   /££££££
-    /££/   /££__  ££ /££__  ££ |____  ££
-   /££/   | ££  \\ ££| ££  \\ ££  /£££££££
-  /££/    | ££  | ££| ££  | ££ /££__  ££
- /££££££££|  ££££££/| £££££££/|  £££££££
-|________/ \\______/ | ££____/  \\_______/
-                    | ££
-                    | ££          React Components
-                    |__/`)}
-
-${kleur.italic('You can now view your style guide in the browser')}
-
-${kleur.bold('Local:')}            ${urls.localUrlForTerminal}
-    `);
-
-  if (urls.lanUrlForTerminal) {
-    console.log(`${kleur.bold('On your network:')}  ${urls.lanUrlForTerminal}
-      `);
-  }
-};
 
 module.exports = {
   title: 'React components',
